convert eval data to float

+++ PointPWC/train_self.py
@@ -223,11 +223,11 @@ def eval_sceneflow(model, loader):
         pos1, pos2, norm1, norm2, flow, _ = data  
 
         #move to cuda 
-        pos1 = pos1.cuda()
-        pos2 = pos2.cuda()
-        norm1 = norm1.cuda()
-        norm2 = norm2.cuda()
-        flow = flow.cuda()
+        pos1 = pos1.cuda().float()
+        pos2 = pos2.cuda().float()
+        norm1 = norm1.cuda().float()
+        norm2 = norm2.cuda().float()
+        flow = flow.cuda().float()
 
         with torch.no_grad():
             pred_flows, _, _, pc1, pc2 = model(pos1, pos2, norm1, norm2)
